refactor(linter): split command option facts
